services/analysis-engine/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
transformers==4.38.1
torch==2.2.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
scipy==1.12.0
numpy==1.26.4
python-multipart==0.0.9
aiohttp==3.9.3
python-dotenv==1.0.1
